GtkSharp.BlazorWebView.csproj: adjust namespace & add BlazorWebView (Stub)
