+++ src/BlazorWebView/src/GtkSharp.BlazorWebView/GtkSharp.BlazorWebView.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>$(_MauiDotNetTfm)-gtk</TargetFrameworks>
+    <SupportedOSPlatformVersion>3.24</SupportedOSPlatformVersion>
+    <Description>Build Gtk applications with Blazor and WebKit</Description>
+    <DefineConstants>$(DefineConstants);WEBKIT_GTK</DefineConstants>
+    <Nullable>enable</Nullable>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <NoWarn>$(NoWarn);CS1591;RS0041;RS0026;RS0027;RS0016;CS8625;CS8618</NoWarn>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Components.WebView" />
+  </ItemGroup>
+
+</Project>
