+++ src/BlazorWebView/src/GtkSharp.BlazorWebView/GtkSharp.Upstream/FuncLoader/Library.cs
