+++ src/BlazorWebView/src/GtkSharp.BlazorWebView/GtkSharp.Upstream/JavaScriptCore/Value.cs
