+++ src/BlazorWebView/src/GtkSharp.BlazorWebView/GtkSharp.Upstream/WebKit/JavascriptResult.cs
