+++ src/BlazorWebView/src/GtkSharp.BlazorWebView/GtkSharp.Upstream/WebKit/UserScript.cs
