+++ src/BlazorWebView/src/GtkSharp.BlazorWebView/PublicAPI.Unshipped.txt
@@ -1,0 +1,1 @@
+﻿#nullable enable
